BENTO-2294 Create independent timeout controller
