+++ packages/bento-frontend/src/components/Layout/LayoutView.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { withStyles, CssBaseline } from '@material-ui/core';
-import { HashRouter, Route, Switch, useHistory } from 'react-router-dom';
+import { HashRouter, Route, Switch } from 'react-router-dom';
 import aboutPageRoutes from '../../bento/aboutPagesRoutes';
 import Header from '../Header/HeaderView';
 import NavBar from '../NavBar/NavBarContainer';
@@ -31,37 +31,20 @@ import viewUserController from '../../pages/admin/userDetails/viewUserController
 import OverlayWindow from '../OverlayWindow/OverlayWindow';
 import AUTH_MIDDLEWARE_CONFIG from '../Auth/authMiddlewareConfig';
 import CarView from '../../pages/cart/cartController';
-import env from '../../utils/env';
-import { PING_INTERVAL, REDIRECT_AFTER_SIGN_OUT, SHOW_WARNING_BEFORE } from '../../bento/siteWideConfig';
-
-import { AuthenticationMiddlewareGenerator, useAuth } from '@bento-core/authentication';
-import { SessionTimeoutGenerator } from '@bento-core/session-timeout';
+import AuthSessionTimeoutController from '../SessionTimeout/SessionTimeoutController';
+import { AuthenticationMiddlewareGenerator } from '@bento-core/authentication';
 
 import Notifactions from '../Notifications/NotifactionView';
 import DashTemplate from '../../pages/dashTemplate/DashTemplateController';
-import { useGlobal } from '../Global/GlobalProvider';
 
 const ScrollToTop = () => {
   window.scrollTo(0, 0);
   return null;
 };
 
-const { SessionTimeout } = SessionTimeoutGenerator({
-  config: {
-    pingInterval: PING_INTERVAL,
-    thresholdTime: SHOW_WARNING_BEFORE,
-    extendEndpoint: `${env.REACT_APP_AUTH_SERVICE_API}authenticated`,
-    ttlEndpoint: `${env.REACT_APP_AUTH_SERVICE_API}session-ttl`,
-  },
-});
-
 const Layout = ({ classes, isSidebarOpened }) => {
   // Access control imports
   const { LoginRoute, MixedRoute, PrivateRoute, AdminRoute} = AuthenticationMiddlewareGenerator(AUTH_MIDDLEWARE_CONFIG);
-  const { signOut } = useAuth();
-
-  const history = useHistory();
-  const { Notification } = useGlobal();
 
   return (
   <>
@@ -69,10 +52,7 @@ const Layout = ({ classes, isSidebarOpened }) => {
     <HashRouter>
       <>
         <Notifactions />
-        <SessionTimeout
-          signOut={() => signOut(history, REDIRECT_AFTER_SIGN_OUT)}
-          showNotification={(content, duration) => Notification.show(content, duration)}
-        />
+        <AuthSessionTimeoutController />
         <Header />
         <OverlayWindow />
         <NavBar />
